Rename IElementsConfig to singular form
