+++ lib/Models/IElementConfig.ts
